Load music list from file
Haven't tested this and don't know if it works. You should change the js file with the list into a json one just with the array and that would be enough, no need to run js or anything and jquery also deals with parsing the json and all. I don't know if toggleMusic is gonna be called more than once so I just assumed it could be and that it would just return what it already has.