Dual encode for H264 and WebM, emit to respective socket io rooms
Both media formats are encoded and stored in the Publico instance.
At the time chats are pulled from Publico and emitted to the client
they exctract the media format and broadcast only that format to
a socket.io "room" by the same name as the format.

Clients now emit an event, "join", with the format they wish to
receive. The initial chat dump occurs on this join event instead
of connection. Room names at this time are "webm" and "mp4".