Fix mp4 encoding.
The mp4 command had a framerate specified that differed from what
the resulting framerate was by a slight amount (it specified 30fps
but the PTS being set was 10*2.5 => 25fps). This resulted in mp4's
being ever so slightly different in length than webm's, and being
unable to properly roundtrip the mp4's.